Snowflakes are now strings.
Snowflakes are now encoded (and decoded) as strings.

Snowflakes can be sent as either a integer, or a string, but will always be returned as a string.
This is due to some languages not supporting large numbers.
A snowflake will always be a valid integer, so converting it into one is safe.