Removal of `document_ids` column and transactions for new pastes.
Removal of the `document_ids` column (instead usage of `Document::fetch_all(...)`).
It is less DB calls, which should reduce load on the server per request.

Now using transactions in post requests.
This removes the paste and document objects created in the database, when something fails.
A future update will remove failed images as well.