added payload **required** document.
The `POST /pastes` (creating new posts) now takes a payload document.

This document is made so for future iterations where more things are required,
they can be added to the this, instead of having 300 query parameters.

The one note to remember, this document is required,
even if you aren't providing any arguments.