Update timestamps, and fix tests.
There is now three timestamps included in requests.

`timestamp`: This is when the timestamp was first created.
`edited_timestamp`: The is when the paste was most recently edited. (can be null for being never edited.)
`expiry_timestamp`: This is when the paste is going to expire. (can be null for no expiry.)

Tests have also been fixed, as snowflakes are now more cleanly incremented, for easy testing.